test: rework directory structure
move unit tests to `test/unit`
