+++ pytest.ini
@@ -1,3 +1,6 @@
 [pytest]
 pythonpath = .
 addopts = --ignore=./src --no-header -v --cov-report=xml --cov=src
+markers =
+	unit: unit tests
+	integration: integration tests
